+++ src/util.h
@@ -16,7 +16,7 @@
 uint64_t get_time();
 void get_time_str(char *str, size_t str_size);
 
-void split_freq(uint64_t freq, uint16_t *mhz, uint16_t *khz, uint16_t *hz);
+void split_freq(int32_t freq, uint16_t *mhz, uint16_t *khz, uint16_t *hz);
 int32_t align_int(int32_t x, uint16_t step);
 uint64_t align_long(uint64_t x, uint16_t step);
 int32_t limit(int32_t x, int32_t min, int32_t max);
